requirements.txt
Werkzeug==2.0.3
Flask==2.0.3
torch
transformers
gitpython
